Debugged the change in I2CBus
Code compiles and runs now (only tested raw mode yet).
One possible regression is, that the constructor I2CBus(int fd) is now deleted.
This due to the problem to copy file descriptors in a way that ioctl on on does not affect
the copied fd (dup() is not the right choice as I first thought)
It would be possible if necessary, but would be a bit messy.